# swagger.json
{"openapi":"3.0.0","paths":{"/healthcheck":{"get":{"operationId":"HealthcheckController_getStatus","summary":"Get server status","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthcheckDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["healthcheck"]}},"/product/info":{"get":{"operationId":"ProductController_getAllProducts","summary":"Get all products data","parameters":[],"responses":{"200":{"description":"Returns a JSON with all the products data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllProductsResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["product"],"security":[{"auth":[]}]}},"/product/info/{id}":{"get":{"operationId":"ProductController_getProduct","summary":"Get the product data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a JSON with the product data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProductResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["product"],"security":[{"auth":[]}]}},"/product/insert":{"put":{"operationId":"ProductController_putProduct","summary":"Put the product data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutProductReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the product data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProductResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["product"],"security":[{"auth":[]}]}},"/product/remove/{id}":{"delete":{"operationId":"ProductController_deleteProduct","summary":"Delete the product data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a JSON with the product status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteProductResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["product"],"security":[{"auth":[]}]}},"/rating/info":{"get":{"operationId":"RatingController_getAllRating","summary":"Get all rating data","parameters":[],"responses":{"200":{"description":"Returns a JSON with all the rating data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllRatingResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["rating"],"security":[{"auth":[]}]}},"/rating/info/{{id}}":{"get":{"operationId":"RatingController_getRating","summary":"Get the rating data","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"number"}},{"name":"comment","in":"query","required":true,"schema":{"type":"string"}},{"name":"grade","in":"query","required":true,"schema":{"type":"number"}},{"name":"productId","in":"query","required":true,"schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a JSON with the rating data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetRatingResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["rating"],"security":[{"auth":[]}]}},"/rating/insert":{"put":{"operationId":"RatingController_putRating","summary":"Put the rating data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"comment","required":true,"in":"path","schema":{"type":"string"}},{"name":"grade","required":true,"in":"path","schema":{"type":"number"}},{"name":"productId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutRatingReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the rating data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetRatingResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["rating"],"security":[{"auth":[]}]}},"/rating/remove/{id}":{"delete":{"operationId":"RatingController_deleteProduct","summary":"Delete the rating data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Returns a JSON with the rating status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteRatingResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["rating"],"security":[{"auth":[]}]}}},"info":{"title":"pdp-ms-product","description":"cosmetics e-commerce","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/pdp-ms-product"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"HealthcheckDto":{"type":"object","properties":{"name":{"type":"string"},"version":{"type":"string"}},"required":["name","version"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetAllProductsResDto":{"type":"object","properties":{"product":{"type":"array","items":{"type":"string"}}},"required":["product"]},"ProductDto":{"type":"object","properties":{"id":{"type":"number"},"brand":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"en_name":{"type":"string"},"pt_name":{"type":"string"},"en_type":{"type":"string"},"pt_type":{"type":"string"},"en_desc":{"type":"string"},"pt_desc":{"type":"string"}},"required":["id","brand","price","quantity","en_name","pt_name","en_type","pt_type","en_desc","pt_desc"]},"GetProductResDto":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductDto"}},"required":["product"]},"PutProductReqDto":{"type":"object","properties":{"id":{"type":"number"},"brand":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"en_name":{"type":"string"},"pt_name":{"type":"string"},"en_type":{"type":"string"},"pt_type":{"type":"string"},"en_desc":{"type":"string"},"pt_desc":{"type":"string"}},"required":["id","brand","price","quantity","en_name","pt_name","en_type","pt_type","en_desc","pt_desc"]},"DeleteProductResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetAllRatingResDto":{"type":"object","properties":{"rating":{"type":"array","items":{"type":"string"}}},"required":["rating"]},"RatingDto":{"type":"object","properties":{"id":{"type":"number"},"comment":{"type":"string"},"grade":{"type":"number"},"productId":{"type":"number"}},"required":["id","comment","grade","productId"]},"GetRatingResDto":{"type":"object","properties":{"rating":{"$ref":"#/components/schemas/RatingDto"}},"required":["rating"]},"PutRatingReqDto":{"type":"object","properties":{"id":{"type":"number"},"comment":{"type":"string"},"grade":{"type":"number"},"productId":{"type":"number"}},"required":["id","comment","grade","productId"]},"DeleteRatingResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}